// src/permissions/permissions.enum.ts
/*
 * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
 *
 * SPDX-License-Identifier: AGPL-3.0-only
 */

/*
 * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
 *
 * SPDX-License-Identifier: AGPL-3.0-only
 */

/**
 * Represents the Permissions a user may hold in a request
 */
export enum Permission {
  READ = 'read',
  WRITE = 'write',
  CREATE = 'create',
  OWNER = 'owner',
}